requirements.txt
fastapi
uvicorn
python-dotenv
duckduckgo-search
requests